refactor(borrow-dialog): move into pages folder
